deps/v8/test/webkit/dfg-integer-optimization-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This tests that integer addition optimizations in the DFG are not performed too overzealously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.a is 1
PASS result.b is 1
PASS result.a is 20000001
PASS result.b is 20000001
PASS result.a is 40000001
PASS result.b is 40000001
PASS result.a is 60000001
PASS result.b is 60000001
PASS result.a is 80000001
PASS result.b is 80000001
PASS result.a is 100000001
PASS result.b is 100000001
PASS result.a is 120000001
PASS result.b is 120000001
PASS result.a is 140000001
PASS result.b is 140000001
PASS result.a is 160000001
PASS result.b is 160000001
PASS result.a is 180000001
PASS result.b is 180000001
PASS result.a is 200000001
PASS result.b is 200000001
PASS result.a is 220000001
PASS result.b is 220000001
PASS result.a is 240000001
PASS result.b is 240000001
PASS result.a is 260000001
PASS result.b is 260000001
PASS result.a is 280000001
PASS result.b is 280000001
PASS result.a is 300000001
PASS result.b is 300000001
PASS result.a is 320000001
PASS result.b is 320000001
PASS result.a is 340000001
PASS result.b is 340000001
PASS result.a is 360000001
PASS result.b is 360000001
PASS result.a is 380000001
PASS result.b is 380000001
PASS result.a is 400000001
PASS result.b is 400000001
PASS result.a is 420000001
PASS result.b is 420000001
PASS result.a is 440000001
PASS result.b is 440000001
PASS result.a is 460000001
PASS result.b is 460000001
PASS result.a is 480000001
PASS result.b is 480000001
PASS result.a is 500000001
PASS result.b is 500000001
PASS result.a is 520000001
PASS result.b is 520000001
PASS result.a is 540000001
PASS result.b is 540000001
PASS result.a is 560000001
PASS result.b is 560000001
PASS result.a is 580000001
PASS result.b is 580000001
PASS result.a is 600000001
PASS result.b is 600000001
PASS result.a is 620000001
PASS result.b is 620000001
PASS result.a is 640000001
PASS result.b is 640000001
PASS result.a is 660000001
PASS result.b is 660000001
PASS result.a is 680000001
PASS result.b is 680000001
PASS result.a is 700000001
PASS result.b is 700000001
PASS result.a is 720000001
PASS result.b is 720000001
PASS result.a is 740000001
PASS result.b is 740000001
PASS result.a is 760000001
PASS result.b is 760000001
PASS result.a is 780000001
PASS result.b is 780000001
PASS result.a is 800000001
PASS result.b is 800000001
PASS result.a is 820000001
PASS result.b is 820000001
PASS result.a is 840000001
PASS result.b is 840000001
PASS result.a is 860000001
PASS result.b is 860000001
PASS result.a is 880000001
PASS result.b is 880000001
PASS result.a is 900000001
PASS result.b is 900000001
PASS result.a is 920000001
PASS result.b is 920000001
PASS result.a is 940000001
PASS result.b is 940000001
PASS result.a is 960000001
PASS result.b is 960000001
PASS result.a is 980000001
PASS result.b is 980000001
PASS result.a is 1000000001
PASS result.b is 1000000001
PASS result.a is 1020000001
PASS result.b is 1020000001
PASS result.a is 1040000001
PASS result.b is 1040000001
PASS result.a is 1060000001
PASS result.b is 1060000001
PASS result.a is 1080000001
PASS result.b is 1080000001
PASS result.a is 1100000001
PASS result.b is 1100000001
PASS result.a is 1120000001
PASS result.b is 1120000001
PASS result.a is 1140000001
PASS result.b is 1140000001
PASS result.a is 1160000001
PASS result.b is 1160000001
PASS result.a is 1180000001
PASS result.b is 1180000001
PASS result.a is 1200000001
PASS result.b is 1200000001
PASS result.a is 1220000001
PASS result.b is 1220000001
PASS result.a is 1240000001
PASS result.b is 1240000001
PASS result.a is 1260000001
PASS result.b is 1260000001
PASS result.a is 1280000001
PASS result.b is 1280000001
PASS result.a is 1300000001
PASS result.b is 1300000001
PASS result.a is 1320000001
PASS result.b is 1320000001
PASS result.a is 1340000001
PASS result.b is 1340000001
PASS result.a is 1360000001
PASS result.b is 1360000001
PASS result.a is 1380000001
PASS result.b is 1380000001
PASS result.a is 1400000001
PASS result.b is 1400000001
PASS result.a is 1420000001
PASS result.b is 1420000001
PASS result.a is 1440000001
PASS result.b is 1440000001
PASS result.a is 1460000001
PASS result.b is 1460000001
PASS result.a is 1480000001
PASS result.b is 1480000001
PASS result.a is 1500000001
PASS result.b is 1500000001
PASS result.a is 1520000001
PASS result.b is 1520000001
PASS result.a is 1540000001
PASS result.b is 1540000001
PASS result.a is 1560000001
PASS result.b is 1560000001
PASS result.a is 1580000001
PASS result.b is 1580000001
PASS result.a is 1600000001
PASS result.b is 1600000001
PASS result.a is 1620000001
PASS result.b is 1620000001
PASS result.a is 1640000001
PASS result.b is 1640000001
PASS result.a is 1660000001
PASS result.b is 1660000001
PASS result.a is 1680000001
PASS result.b is 1680000001
PASS result.a is 1700000001
PASS result.b is 1700000001
PASS result.a is 1720000001
PASS result.b is 1720000001
PASS result.a is 1740000001
PASS result.b is 1740000001
PASS result.a is 1760000001
PASS result.b is 1760000001
PASS result.a is 1780000001
PASS result.b is 1780000001
PASS result.a is 1800000001
PASS result.b is 1800000001
PASS result.a is 1820000001
PASS result.b is 1820000001
PASS result.a is 1840000001
PASS result.b is 1840000001
PASS result.a is 1860000001
PASS result.b is 1860000001
PASS result.a is 1880000001
PASS result.b is 1880000001
PASS result.a is 1900000001
PASS result.b is 1900000001
PASS result.a is 1920000001
PASS result.b is 1920000001
PASS result.a is 1940000001
PASS result.b is 1940000001
PASS result.a is 1960000001
PASS result.b is 1960000001
PASS result.a is 1980000001
PASS result.b is 1980000001
PASS result.a is 2000000001
PASS result.b is 2000000001
PASS result.a is 2020000001
PASS result.b is 2020000001
PASS result.a is 2040000001
PASS result.b is 2040000001
PASS result.a is 2060000001
PASS result.b is 2060000001
PASS result.a is 2080000001
PASS result.b is 2080000001
PASS result.a is 2100000001
PASS result.b is 2100000001
PASS result.a is 2120000001
PASS result.b is 2120000001
PASS result.a is 2140000001
PASS result.b is 2140000001
PASS result.a is -2134967295
PASS result.b is 2160000001
PASS result.a is -2114967295
PASS result.b is 2180000001
PASS result.a is -2094967295
PASS result.b is 2200000001
PASS result.a is -2074967295
PASS result.b is 2220000001
PASS result.a is -2054967295
PASS result.b is 2240000001
PASS result.a is -2034967295
PASS result.b is 2260000001
PASS result.a is -2014967295
PASS result.b is 2280000001
PASS result.a is -1994967295
PASS result.b is 2300000001
PASS result.a is -1974967295
PASS result.b is 2320000001
PASS result.a is -1954967295
PASS result.b is 2340000001
PASS result.a is -1934967295
PASS result.b is 2360000001
PASS result.a is -1914967295
PASS result.b is 2380000001
PASS result.a is -1894967295
PASS result.b is 2400000001
PASS result.a is -1874967295
PASS result.b is 2420000001
PASS result.a is -1854967295
PASS result.b is 2440000001
PASS result.a is -1834967295
PASS result.b is 2460000001
PASS result.a is -1814967295
PASS result.b is 2480000001
PASS result.a is -1794967295
PASS result.b is 2500000001
PASS result.a is -1774967295
PASS result.b is 2520000001
PASS result.a is -1754967295
PASS result.b is 2540000001
PASS result.a is -1734967295
PASS result.b is 2560000001
PASS result.a is -1714967295
PASS result.b is 2580000001
PASS result.a is -1694967295
PASS result.b is 2600000001
PASS result.a is -1674967295
PASS result.b is 2620000001
PASS result.a is -1654967295
PASS result.b is 2640000001
PASS result.a is -1634967295
PASS result.b is 2660000001
PASS result.a is -1614967295
PASS result.b is 2680000001
PASS result.a is -1594967295
PASS result.b is 2700000001
PASS result.a is -1574967295
PASS result.b is 2720000001
PASS result.a is -1554967295
PASS result.b is 2740000001
PASS result.a is -1534967295
PASS result.b is 2760000001
PASS result.a is -1514967295
PASS result.b is 2780000001
PASS result.a is -1494967295
PASS result.b is 2800000001
PASS result.a is -1474967295
PASS result.b is 2820000001
PASS result.a is -1454967295
PASS result.b is 2840000001
PASS result.a is -1434967295
PASS result.b is 2860000001
PASS result.a is -1414967295
PASS result.b is 2880000001
PASS result.a is -1394967295
PASS result.b is 2900000001
PASS result.a is -1374967295
PASS result.b is 2920000001
PASS result.a is -1354967295
PASS result.b is 2940000001
PASS result.a is -1334967295
PASS result.b is 2960000001
PASS result.a is -1314967295
PASS result.b is 2980000001
PASS result.a is -1294967295
PASS result.b is 3000000001
PASS result.a is -1274967295
PASS result.b is 3020000001
PASS result.a is -1254967295
PASS result.b is 3040000001
PASS result.a is -1234967295
PASS result.b is 3060000001
PASS result.a is -1214967295
PASS result.b is 3080000001
PASS result.a is -1194967295
PASS result.b is 3100000001
PASS result.a is -1174967295
PASS result.b is 3120000001
PASS result.a is -1154967295
PASS result.b is 3140000001
PASS result.a is -1134967295
PASS result.b is 3160000001
PASS result.a is -1114967295
PASS result.b is 3180000001
PASS result.a is -1094967295
PASS result.b is 3200000001
PASS result.a is -1074967295
PASS result.b is 3220000001
PASS result.a is -1054967295
PASS result.b is 3240000001
PASS result.a is -1034967295
PASS result.b is 3260000001
PASS result.a is -1014967295
PASS result.b is 3280000001
PASS result.a is -994967295
PASS result.b is 3300000001
PASS result.a is -974967295
PASS result.b is 3320000001
PASS result.a is -954967295
PASS result.b is 3340000001
PASS result.a is -934967295
PASS result.b is 3360000001
PASS result.a is -914967295
PASS result.b is 3380000001
PASS result.a is -894967295
PASS result.b is 3400000001
PASS result.a is -874967295
PASS result.b is 3420000001
PASS result.a is -854967295
PASS result.b is 3440000001
PASS result.a is -834967295
PASS result.b is 3460000001
PASS result.a is -814967295
PASS result.b is 3480000001
PASS result.a is -794967295
PASS result.b is 3500000001
PASS result.a is -774967295
PASS result.b is 3520000001
PASS result.a is -754967295
PASS result.b is 3540000001
PASS result.a is -734967295
PASS result.b is 3560000001
PASS result.a is -714967295
PASS result.b is 3580000001
PASS result.a is -694967295
PASS result.b is 3600000001
PASS result.a is -674967295
PASS result.b is 3620000001
PASS result.a is -654967295
PASS result.b is 3640000001
PASS result.a is -634967295
PASS result.b is 3660000001
PASS result.a is -614967295
PASS result.b is 3680000001
PASS result.a is -594967295
PASS result.b is 3700000001
PASS result.a is -574967295
PASS result.b is 3720000001
PASS result.a is -554967295
PASS result.b is 3740000001
PASS result.a is -534967295
PASS result.b is 3760000001
PASS result.a is -514967295
PASS result.b is 3780000001
PASS result.a is -494967295
PASS result.b is 3800000001
PASS result.a is -474967295
PASS result.b is 3820000001
PASS result.a is -454967295
PASS result.b is 3840000001
PASS result.a is -434967295
PASS result.b is 3860000001
PASS result.a is -414967295
PASS result.b is 3880000001
PASS result.a is -394967295
PASS result.b is 3900000001
PASS result.a is -374967295
PASS result.b is 3920000001
PASS result.a is -354967295
PASS result.b is 3940000001
PASS result.a is -334967295
PASS result.b is 3960000001
PASS result.a is -314967295
PASS result.b is 3980000001
PASS result.a is -294967295
PASS result.b is 4000000001
PASS result.a is -274967295
PASS result.b is 4020000001
PASS result.a is -254967295
PASS result.b is 4040000001
PASS result.a is -234967295
PASS result.b is 4060000001
PASS result.a is -214967295
PASS result.b is 4080000001
PASS result.a is -194967295
PASS result.b is 4100000001
PASS result.a is -174967295
PASS result.b is 4120000001
PASS result.a is -154967295
PASS result.b is 4140000001
PASS result.a is -134967295
PASS result.b is 4160000001
PASS result.a is -114967295
PASS result.b is 4180000001
PASS result.a is -94967295
PASS result.b is 4200000001
PASS result.a is -74967295
PASS result.b is 4220000001
PASS result.a is -54967295
PASS result.b is 4240000001
PASS result.a is -34967295
PASS result.b is 4260000001
PASS result.a is -14967295
PASS result.b is 4280000001
PASS result.a is 5032705
PASS result.b is 4300000001
PASS result.a is 25032705
PASS result.b is 4320000001
PASS result.a is 45032705
PASS result.b is 4340000001
PASS result.a is 65032705
PASS result.b is 4360000001
PASS result.a is 85032705
PASS result.b is 4380000001
PASS result.a is 105032705
PASS result.b is 4400000001
PASS result.a is 125032705
PASS result.b is 4420000001
PASS result.a is 145032705
PASS result.b is 4440000001
PASS result.a is 165032705
PASS result.b is 4460000001
PASS result.a is 185032705
PASS result.b is 4480000001
PASS result.a is 205032705
PASS result.b is 4500000001
PASS result.a is 225032705
PASS result.b is 4520000001
PASS result.a is 245032705
PASS result.b is 4540000001
PASS result.a is 265032705
PASS result.b is 4560000001
PASS result.a is 285032705
PASS result.b is 4580000001
PASS result.a is 305032705
PASS result.b is 4600000001
PASS result.a is 325032705
PASS result.b is 4620000001
PASS result.a is 345032705
PASS result.b is 4640000001
PASS result.a is 365032705
PASS result.b is 4660000001
PASS result.a is 385032705
PASS result.b is 4680000001
PASS result.a is 405032705
PASS result.b is 4700000001
PASS result.a is 425032705
PASS result.b is 4720000001
PASS result.a is 445032705
PASS result.b is 4740000001
PASS result.a is 465032705
PASS result.b is 4760000001
PASS result.a is 485032705
PASS result.b is 4780000001
PASS result.a is 505032705
PASS result.b is 4800000001
PASS result.a is 525032705
PASS result.b is 4820000001
PASS result.a is 545032705
PASS result.b is 4840000001
PASS result.a is 565032705
PASS result.b is 4860000001
PASS result.a is 585032705
PASS result.b is 4880000001
PASS result.a is 605032705
PASS result.b is 4900000001
PASS result.a is 625032705
PASS result.b is 4920000001
PASS result.a is 645032705
PASS result.b is 4940000001
PASS result.a is 665032705
PASS result.b is 4960000001
PASS result.a is 685032705
PASS result.b is 4980000001
PASS result.a is 705032705
PASS result.b is 5000000001
PASS result.a is 725032705
PASS result.b is 5020000001
PASS result.a is 745032705
PASS result.b is 5040000001
PASS result.a is 765032705
PASS result.b is 5060000001
PASS result.a is 785032705
PASS result.b is 5080000001
PASS result.a is 805032705
PASS result.b is 5100000001
PASS result.a is 825032705
PASS result.b is 5120000001
PASS result.a is 845032705
PASS result.b is 5140000001
PASS result.a is 865032705
PASS result.b is 5160000001
PASS result.a is 885032705
PASS result.b is 5180000001
PASS result.a is 905032705
PASS result.b is 5200000001
PASS result.a is 925032705
PASS result.b is 5220000001
PASS result.a is 945032705
PASS result.b is 5240000001
PASS result.a is 965032705
PASS result.b is 5260000001
PASS result.a is 985032705
PASS result.b is 5280000001
PASS result.a is 1005032705
PASS result.b is 5300000001
PASS result.a is 1025032705
PASS result.b is 5320000001
PASS result.a is 1045032705
PASS result.b is 5340000001
PASS result.a is 1065032705
PASS result.b is 5360000001
PASS result.a is 1085032705
PASS result.b is 5380000001
PASS result.a is 1105032705
PASS result.b is 5400000001
PASS result.a is 1125032705
PASS result.b is 5420000001
PASS result.a is 1145032705
PASS result.b is 5440000001
PASS result.a is 1165032705
PASS result.b is 5460000001
PASS result.a is 1185032705
PASS result.b is 5480000001
PASS result.a is 1205032705
PASS result.b is 5500000001
PASS result.a is 1225032705
PASS result.b is 5520000001
PASS result.a is 1245032705
PASS result.b is 5540000001
PASS result.a is 1265032705
PASS result.b is 5560000001
PASS result.a is 1285032705
PASS result.b is 5580000001
PASS result.a is 1305032705
PASS result.b is 5600000001
PASS result.a is 1325032705
PASS result.b is 5620000001
PASS result.a is 1345032705
PASS result.b is 5640000001
PASS result.a is 1365032705
PASS result.b is 5660000001
PASS result.a is 1385032705
PASS result.b is 5680000001
PASS result.a is 1405032705
PASS result.b is 5700000001
PASS result.a is 1425032705
PASS result.b is 5720000001
PASS result.a is 1445032705
PASS result.b is 5740000001
PASS result.a is 1465032705
PASS result.b is 5760000001
PASS result.a is 1485032705
PASS result.b is 5780000001
PASS result.a is 1505032705
PASS result.b is 5800000001
PASS result.a is 1525032705
PASS result.b is 5820000001
PASS result.a is 1545032705
PASS result.b is 5840000001
PASS result.a is 1565032705
PASS result.b is 5860000001
PASS result.a is 1585032705
PASS result.b is 5880000001
PASS result.a is 1605032705
PASS result.b is 5900000001
PASS result.a is 1625032705
PASS result.b is 5920000001
PASS result.a is 1645032705
PASS result.b is 5940000001
PASS result.a is 1665032705
PASS result.b is 5960000001
PASS result.a is 1685032705
PASS result.b is 5980000001
PASS result.a is 1705032705
PASS result.b is 6000000001
PASS result.a is 1725032705
PASS result.b is 6020000001
PASS result.a is 1745032705
PASS result.b is 6040000001
PASS result.a is 1765032705
PASS result.b is 6060000001
PASS result.a is 1785032705
PASS result.b is 6080000001
PASS result.a is 1805032705
PASS result.b is 6100000001
PASS result.a is 1825032705
PASS result.b is 6120000001
PASS result.a is 1845032705
PASS result.b is 6140000001
PASS result.a is 1865032705
PASS result.b is 6160000001
PASS result.a is 1885032705
PASS result.b is 6180000001
PASS result.a is 1905032705
PASS result.b is 6200000001
PASS result.a is 1925032705
PASS result.b is 6220000001
PASS result.a is 1945032705
PASS result.b is 6240000001
PASS result.a is 1965032705
PASS result.b is 6260000001
PASS result.a is 1985032705
PASS result.b is 6280000001
PASS result.a is 2005032705
PASS result.b is 6300000001
PASS result.a is 2025032705
PASS result.b is 6320000001
PASS result.a is 2045032705
PASS result.b is 6340000001
PASS result.a is 2065032705
PASS result.b is 6360000001
PASS result.a is 2085032705
PASS result.b is 6380000001
PASS result.a is 2105032705
PASS result.b is 6400000001
PASS result.a is 2125032705
PASS result.b is 6420000001
PASS result.a is 2145032705
PASS result.b is 6440000001
PASS result.a is -2129934591
PASS result.b is 6460000001
PASS result.a is -2109934591
PASS result.b is 6480000001
PASS result.a is -2089934591
PASS result.b is 6500000001
PASS result.a is -2069934591
PASS result.b is 6520000001
PASS result.a is -2049934591
PASS result.b is 6540000001
PASS result.a is -2029934591
PASS result.b is 6560000001
PASS result.a is -2009934591
PASS result.b is 6580000001
PASS result.a is -1989934591
PASS result.b is 6600000001
PASS result.a is -1969934591
PASS result.b is 6620000001
PASS result.a is -1949934591
PASS result.b is 6640000001
PASS result.a is -1929934591
PASS result.b is 6660000001
PASS result.a is -1909934591
PASS result.b is 6680000001
PASS result.a is -1889934591
PASS result.b is 6700000001
PASS result.a is -1869934591
PASS result.b is 6720000001
PASS result.a is -1849934591
PASS result.b is 6740000001
PASS result.a is -1829934591
PASS result.b is 6760000001
PASS result.a is -1809934591
PASS result.b is 6780000001
PASS result.a is -1789934591
PASS result.b is 6800000001
PASS result.a is -1769934591
PASS result.b is 6820000001
PASS result.a is -1749934591
PASS result.b is 6840000001
PASS result.a is -1729934591
PASS result.b is 6860000001
PASS result.a is -1709934591
PASS result.b is 6880000001
PASS result.a is -1689934591
PASS result.b is 6900000001
PASS result.a is -1669934591
PASS result.b is 6920000001
PASS result.a is -1649934591
PASS result.b is 6940000001
PASS result.a is -1629934591
PASS result.b is 6960000001
PASS result.a is -1609934591
PASS result.b is 6980000001
PASS result.a is -1589934591
PASS result.b is 7000000001
PASS result.a is -1569934591
PASS result.b is 7020000001
PASS result.a is -1549934591
PASS result.b is 7040000001
PASS result.a is -1529934591
PASS result.b is 7060000001
PASS result.a is -1509934591
PASS result.b is 7080000001
PASS result.a is -1489934591
PASS result.b is 7100000001
PASS result.a is -1469934591
PASS result.b is 7120000001
PASS result.a is -1449934591
PASS result.b is 7140000001
PASS result.a is -1429934591
PASS result.b is 7160000001
PASS result.a is -1409934591
PASS result.b is 7180000001
PASS result.a is -1389934591
PASS result.b is 7200000001
PASS result.a is -1369934591
PASS result.b is 7220000001
PASS result.a is -1349934591
PASS result.b is 7240000001
PASS result.a is -1329934591
PASS result.b is 7260000001
PASS result.a is -1309934591
PASS result.b is 7280000001
PASS result.a is -1289934591
PASS result.b is 7300000001
PASS result.a is -1269934591
PASS result.b is 7320000001
PASS result.a is -1249934591
PASS result.b is 7340000001
PASS result.a is -1229934591
PASS result.b is 7360000001
PASS result.a is -1209934591
PASS result.b is 7380000001
PASS result.a is -1189934591
PASS result.b is 7400000001
PASS result.a is -1169934591
PASS result.b is 7420000001
PASS result.a is -1149934591
PASS result.b is 7440000001
PASS result.a is -1129934591
PASS result.b is 7460000001
PASS result.a is -1109934591
PASS result.b is 7480000001
PASS result.a is -1089934591
PASS result.b is 7500000001
PASS result.a is -1069934591
PASS result.b is 7520000001
PASS result.a is -1049934591
PASS result.b is 7540000001
PASS result.a is -1029934591
PASS result.b is 7560000001
PASS result.a is -1009934591
PASS result.b is 7580000001
PASS result.a is -989934591
PASS result.b is 7600000001
PASS result.a is -969934591
PASS result.b is 7620000001
PASS result.a is -949934591
PASS result.b is 7640000001
PASS result.a is -929934591
PASS result.b is 7660000001
PASS result.a is -909934591
PASS result.b is 7680000001
PASS result.a is -889934591
PASS result.b is 7700000001
PASS result.a is -869934591
PASS result.b is 7720000001
PASS result.a is -849934591
PASS result.b is 7740000001
PASS result.a is -829934591
PASS result.b is 7760000001
PASS result.a is -809934591
PASS result.b is 7780000001
PASS result.a is -789934591
PASS result.b is 7800000001
PASS result.a is -769934591
PASS result.b is 7820000001
PASS result.a is -749934591
PASS result.b is 7840000001
PASS result.a is -729934591
PASS result.b is 7860000001
PASS result.a is -709934591
PASS result.b is 7880000001
PASS result.a is -689934591
PASS result.b is 7900000001
PASS result.a is -669934591
PASS result.b is 7920000001
PASS result.a is -649934591
PASS result.b is 7940000001
PASS result.a is -629934591
PASS result.b is 7960000001
PASS result.a is -609934591
PASS result.b is 7980000001
PASS result.a is -589934591
PASS result.b is 8000000001
PASS result.a is -569934591
PASS result.b is 8020000001
PASS result.a is -549934591
PASS result.b is 8040000001
PASS result.a is -529934591
PASS result.b is 8060000001
PASS result.a is -509934591
PASS result.b is 8080000001
PASS result.a is -489934591
PASS result.b is 8100000001
PASS result.a is -469934591
PASS result.b is 8120000001
PASS result.a is -449934591
PASS result.b is 8140000001
PASS result.a is -429934591
PASS result.b is 8160000001
PASS result.a is -409934591
PASS result.b is 8180000001
PASS result.a is -389934591
PASS result.b is 8200000001
PASS result.a is -369934591
PASS result.b is 8220000001
PASS result.a is -349934591
PASS result.b is 8240000001
PASS result.a is -329934591
PASS result.b is 8260000001
PASS result.a is -309934591
PASS result.b is 8280000001
PASS result.a is -289934591
PASS result.b is 8300000001
PASS result.a is -269934591
PASS result.b is 8320000001
PASS result.a is -249934591
PASS result.b is 8340000001
PASS result.a is -229934591
PASS result.b is 8360000001
PASS result.a is -209934591
PASS result.b is 8380000001
PASS result.a is -189934591
PASS result.b is 8400000001
PASS result.a is -169934591
PASS result.b is 8420000001
PASS result.a is -149934591
PASS result.b is 8440000001
PASS result.a is -129934591
PASS result.b is 8460000001
PASS result.a is -109934591
PASS result.b is 8480000001
PASS result.a is -89934591
PASS result.b is 8500000001
PASS result.a is -69934591
PASS result.b is 8520000001
PASS result.a is -49934591
PASS result.b is 8540000001
PASS result.a is -29934591
PASS result.b is 8560000001
PASS result.a is -9934591
PASS result.b is 8580000001
PASS result.a is 10065409
PASS result.b is 8600000001
PASS result.a is 30065409
PASS result.b is 8620000001
PASS result.a is 50065409
PASS result.b is 8640000001
PASS result.a is 70065409
PASS result.b is 8660000001
PASS result.a is 90065409
PASS result.b is 8680000001
PASS result.a is 110065409
PASS result.b is 8700000001
PASS result.a is 130065409
PASS result.b is 8720000001
PASS result.a is 150065409
PASS result.b is 8740000001
PASS result.a is 170065409
PASS result.b is 8760000001
PASS result.a is 190065409
PASS result.b is 8780000001
PASS result.a is 210065409
PASS result.b is 8800000001
PASS result.a is 230065409
PASS result.b is 8820000001
PASS result.a is 250065409
PASS result.b is 8840000001
PASS result.a is 270065409
PASS result.b is 8860000001
PASS result.a is 290065409
PASS result.b is 8880000001
PASS result.a is 310065409
PASS result.b is 8900000001
PASS result.a is 330065409
PASS result.b is 8920000001
PASS result.a is 350065409
PASS result.b is 8940000001
PASS result.a is 370065409
PASS result.b is 8960000001
PASS result.a is 390065409
PASS result.b is 8980000001
PASS result.a is 410065409
PASS result.b is 9000000001
PASS result.a is 430065409
PASS result.b is 9020000001
PASS result.a is 450065409
PASS result.b is 9040000001
PASS result.a is 470065409
PASS result.b is 9060000001
PASS result.a is 490065409
PASS result.b is 9080000001
PASS result.a is 510065409
PASS result.b is 9100000001
PASS result.a is 530065409
PASS result.b is 9120000001
PASS result.a is 550065409
PASS result.b is 9140000001
PASS result.a is 570065409
PASS result.b is 9160000001
PASS result.a is 590065409
PASS result.b is 9180000001
PASS result.a is 610065409
PASS result.b is 9200000001
PASS result.a is 630065409
PASS result.b is 9220000001
PASS result.a is 650065409
PASS result.b is 9240000001
PASS result.a is 670065409
PASS result.b is 9260000001
PASS result.a is 690065409
PASS result.b is 9280000001
PASS result.a is 710065409
PASS result.b is 9300000001
PASS result.a is 730065409
PASS result.b is 9320000001
PASS result.a is 750065409
PASS result.b is 9340000001
PASS result.a is 770065409
PASS result.b is 9360000001
PASS result.a is 790065409
PASS result.b is 9380000001
PASS result.a is 810065409
PASS result.b is 9400000001
PASS result.a is 830065409
PASS result.b is 9420000001
PASS result.a is 850065409
PASS result.b is 9440000001
PASS result.a is 870065409
PASS result.b is 9460000001
PASS result.a is 890065409
PASS result.b is 9480000001
PASS result.a is 910065409
PASS result.b is 9500000001
PASS result.a is 930065409
PASS result.b is 9520000001
PASS result.a is 950065409
PASS result.b is 9540000001
PASS result.a is 970065409
PASS result.b is 9560000001
PASS result.a is 990065409
PASS result.b is 9580000001
PASS result.a is 1010065409
PASS result.b is 9600000001
PASS result.a is 1030065409
PASS result.b is 9620000001
PASS result.a is 1050065409
PASS result.b is 9640000001
PASS result.a is 1070065409
PASS result.b is 9660000001
PASS result.a is 1090065409
PASS result.b is 9680000001
PASS result.a is 1110065409
PASS result.b is 9700000001
PASS result.a is 1130065409
PASS result.b is 9720000001
PASS result.a is 1150065409
PASS result.b is 9740000001
PASS result.a is 1170065409
PASS result.b is 9760000001
PASS result.a is 1190065409
PASS result.b is 9780000001
PASS result.a is 1210065409
PASS result.b is 9800000001
PASS result.a is 1230065409
PASS result.b is 9820000001
PASS result.a is 1250065409
PASS result.b is 9840000001
PASS result.a is 1270065409
PASS result.b is 9860000001
PASS result.a is 1290065409
PASS result.b is 9880000001
PASS result.a is 1310065409
PASS result.b is 9900000001
PASS result.a is 1330065409
PASS result.b is 9920000001
PASS result.a is 1350065409
PASS result.b is 9940000001
PASS result.a is 1370065409
PASS result.b is 9960000001
PASS result.a is 1390065409
PASS result.b is 9980000001
PASS result.a is 1410065409
PASS result.b is 10000000001
PASS result.a is 1430065409
PASS result.b is 10020000001
PASS result.a is 1450065409
PASS result.b is 10040000001
PASS result.a is 1470065409
PASS result.b is 10060000001
PASS result.a is 1490065409
PASS result.b is 10080000001
PASS result.a is 1510065409
PASS result.b is 10100000001
PASS result.a is 1530065409
PASS result.b is 10120000001
PASS result.a is 1550065409
PASS result.b is 10140000001
PASS result.a is 1570065409
PASS result.b is 10160000001
PASS result.a is 1590065409
PASS result.b is 10180000001
PASS result.a is 1610065409
PASS result.b is 10200000001
PASS result.a is 1630065409
PASS result.b is 10220000001
PASS result.a is 1650065409
PASS result.b is 10240000001
PASS result.a is 1670065409
PASS result.b is 10260000001
PASS result.a is 1690065409
PASS result.b is 10280000001
PASS result.a is 1710065409
PASS result.b is 10300000001
PASS result.a is 1730065409
PASS result.b is 10320000001
PASS result.a is 1750065409
PASS result.b is 10340000001
PASS result.a is 1770065409
PASS result.b is 10360000001
PASS result.a is 1790065409
PASS result.b is 10380000001
PASS result.a is 1810065409
PASS result.b is 10400000001
PASS result.a is 1830065409
PASS result.b is 10420000001
PASS result.a is 1850065409
PASS result.b is 10440000001
PASS result.a is 1870065409
PASS result.b is 10460000001
PASS result.a is 1890065409
PASS result.b is 10480000001
PASS result.a is 1910065409
PASS result.b is 10500000001
PASS result.a is 1930065409
PASS result.b is 10520000001
PASS result.a is 1950065409
PASS result.b is 10540000001
PASS result.a is 1970065409
PASS result.b is 10560000001
PASS result.a is 1990065409
PASS result.b is 10580000001
PASS result.a is 2010065409
PASS result.b is 10600000001
PASS result.a is 2030065409
PASS result.b is 10620000001
PASS result.a is 2050065409
PASS result.b is 10640000001
PASS result.a is 2070065409
PASS result.b is 10660000001
PASS result.a is 2090065409
PASS result.b is 10680000001
PASS result.a is 2110065409
PASS result.b is 10700000001
PASS result.a is 2130065409
PASS result.b is 10720000001
PASS result.a is -2144901887
PASS result.b is 10740000001
PASS result.a is -2124901887
PASS result.b is 10760000001
PASS result.a is -2104901887
PASS result.b is 10780000001
PASS result.a is -2084901887
PASS result.b is 10800000001
PASS result.a is -2064901887
PASS result.b is 10820000001
PASS result.a is -2044901887
PASS result.b is 10840000001
PASS result.a is -2024901887
PASS result.b is 10860000001
PASS result.a is -2004901887
PASS result.b is 10880000001
PASS result.a is -1984901887
PASS result.b is 10900000001
PASS result.a is -1964901887
PASS result.b is 10920000001
PASS result.a is -1944901887
PASS result.b is 10940000001
PASS result.a is -1924901887
PASS result.b is 10960000001
PASS result.a is -1904901887
PASS result.b is 10980000001
PASS result.a is -1884901887
PASS result.b is 11000000001
PASS result.a is -1864901887
PASS result.b is 11020000001
PASS result.a is -1844901887
PASS result.b is 11040000001
PASS result.a is -1824901887
PASS result.b is 11060000001
PASS result.a is -1804901887
PASS result.b is 11080000001
PASS result.a is -1784901887
PASS result.b is 11100000001
PASS result.a is -1764901887
PASS result.b is 11120000001
PASS result.a is -1744901887
PASS result.b is 11140000001
PASS result.a is -1724901887
PASS result.b is 11160000001
PASS result.a is -1704901887
PASS result.b is 11180000001
PASS result.a is -1684901887
PASS result.b is 11200000001
PASS result.a is -1664901887
PASS result.b is 11220000001
PASS result.a is -1644901887
PASS result.b is 11240000001
PASS result.a is -1624901887
PASS result.b is 11260000001
PASS result.a is -1604901887
PASS result.b is 11280000001
PASS result.a is -1584901887
PASS result.b is 11300000001
PASS result.a is -1564901887
PASS result.b is 11320000001
PASS result.a is -1544901887
PASS result.b is 11340000001
PASS result.a is -1524901887
PASS result.b is 11360000001
PASS result.a is -1504901887
PASS result.b is 11380000001
PASS result.a is -1484901887
PASS result.b is 11400000001
PASS result.a is -1464901887
PASS result.b is 11420000001
PASS result.a is -1444901887
PASS result.b is 11440000001
PASS result.a is -1424901887
PASS result.b is 11460000001
PASS result.a is -1404901887
PASS result.b is 11480000001
PASS result.a is -1384901887
PASS result.b is 11500000001
PASS result.a is -1364901887
PASS result.b is 11520000001
PASS result.a is -1344901887
PASS result.b is 11540000001
PASS result.a is -1324901887
PASS result.b is 11560000001
PASS result.a is -1304901887
PASS result.b is 11580000001
PASS result.a is -1284901887
PASS result.b is 11600000001
PASS result.a is -1264901887
PASS result.b is 11620000001
PASS result.a is -1244901887
PASS result.b is 11640000001
PASS result.a is -1224901887
PASS result.b is 11660000001
PASS result.a is -1204901887
PASS result.b is 11680000001
PASS result.a is -1184901887
PASS result.b is 11700000001
PASS result.a is -1164901887
PASS result.b is 11720000001
PASS result.a is -1144901887
PASS result.b is 11740000001
PASS result.a is -1124901887
PASS result.b is 11760000001
PASS result.a is -1104901887
PASS result.b is 11780000001
PASS result.a is -1084901887
PASS result.b is 11800000001
PASS result.a is -1064901887
PASS result.b is 11820000001
PASS result.a is -1044901887
PASS result.b is 11840000001
PASS result.a is -1024901887
PASS result.b is 11860000001
PASS result.a is -1004901887
PASS result.b is 11880000001
PASS result.a is -984901887
PASS result.b is 11900000001
PASS result.a is -964901887
PASS result.b is 11920000001
PASS result.a is -944901887
PASS result.b is 11940000001
PASS result.a is -924901887
PASS result.b is 11960000001
PASS result.a is -904901887
PASS result.b is 11980000001
PASS result.a is -884901887
PASS result.b is 12000000001
PASS result.a is -864901887
PASS result.b is 12020000001
PASS result.a is -844901887
PASS result.b is 12040000001
PASS result.a is -824901887
PASS result.b is 12060000001
PASS result.a is -804901887
PASS result.b is 12080000001
PASS result.a is -784901887
PASS result.b is 12100000001
PASS result.a is -764901887
PASS result.b is 12120000001
PASS result.a is -744901887
PASS result.b is 12140000001
PASS result.a is -724901887
PASS result.b is 12160000001
PASS result.a is -704901887
PASS result.b is 12180000001
PASS result.a is -684901887
PASS result.b is 12200000001
PASS result.a is -664901887
PASS result.b is 12220000001
PASS result.a is -644901887
PASS result.b is 12240000001
PASS result.a is -624901887
PASS result.b is 12260000001
PASS result.a is -604901887
PASS result.b is 12280000001
PASS result.a is -584901887
PASS result.b is 12300000001
PASS result.a is -564901887
PASS result.b is 12320000001
PASS result.a is -544901887
PASS result.b is 12340000001
PASS result.a is -524901887
PASS result.b is 12360000001
PASS result.a is -504901887
PASS result.b is 12380000001
PASS result.a is -484901887
PASS result.b is 12400000001
PASS result.a is -464901887
PASS result.b is 12420000001
PASS result.a is -444901887
PASS result.b is 12440000001
PASS result.a is -424901887
PASS result.b is 12460000001
PASS result.a is -404901887
PASS result.b is 12480000001
PASS result.a is -384901887
PASS result.b is 12500000001
PASS result.a is -364901887
PASS result.b is 12520000001
PASS result.a is -344901887
PASS result.b is 12540000001
PASS result.a is -324901887
PASS result.b is 12560000001
PASS result.a is -304901887
PASS result.b is 12580000001
PASS result.a is -284901887
PASS result.b is 12600000001
PASS result.a is -264901887
PASS result.b is 12620000001
PASS result.a is -244901887
PASS result.b is 12640000001
PASS result.a is -224901887
PASS result.b is 12660000001
PASS result.a is -204901887
PASS result.b is 12680000001
PASS result.a is -184901887
PASS result.b is 12700000001
PASS result.a is -164901887
PASS result.b is 12720000001
PASS result.a is -144901887
PASS result.b is 12740000001
PASS result.a is -124901887
PASS result.b is 12760000001
PASS result.a is -104901887
PASS result.b is 12780000001
PASS result.a is -84901887
PASS result.b is 12800000001
PASS result.a is -64901887
PASS result.b is 12820000001
PASS result.a is -44901887
PASS result.b is 12840000001
PASS result.a is -24901887
PASS result.b is 12860000001
PASS result.a is -4901887
PASS result.b is 12880000001
PASS result.a is 15098113
PASS result.b is 12900000001
PASS result.a is 35098113
PASS result.b is 12920000001
PASS result.a is 55098113
PASS result.b is 12940000001
PASS result.a is 75098113
PASS result.b is 12960000001
PASS result.a is 95098113
PASS result.b is 12980000001
PASS result.a is 115098113
PASS result.b is 13000000001
PASS result.a is 135098113
PASS result.b is 13020000001
PASS result.a is 155098113
PASS result.b is 13040000001
PASS result.a is 175098113
PASS result.b is 13060000001
PASS result.a is 195098113
PASS result.b is 13080000001
PASS result.a is 215098113
PASS result.b is 13100000001
PASS result.a is 235098113
PASS result.b is 13120000001
PASS result.a is 255098113
PASS result.b is 13140000001
PASS result.a is 275098113
PASS result.b is 13160000001
PASS result.a is 295098113
PASS result.b is 13180000001
PASS result.a is 315098113
PASS result.b is 13200000001
PASS result.a is 335098113
PASS result.b is 13220000001
PASS result.a is 355098113
PASS result.b is 13240000001
PASS result.a is 375098113
PASS result.b is 13260000001
PASS result.a is 395098113
PASS result.b is 13280000001
PASS result.a is 415098113
PASS result.b is 13300000001
PASS result.a is 435098113
PASS result.b is 13320000001
PASS result.a is 455098113
PASS result.b is 13340000001
PASS result.a is 475098113
PASS result.b is 13360000001
PASS result.a is 495098113
PASS result.b is 13380000001
PASS result.a is 515098113
PASS result.b is 13400000001
PASS result.a is 535098113
PASS result.b is 13420000001
PASS result.a is 555098113
PASS result.b is 13440000001
PASS result.a is 575098113
PASS result.b is 13460000001
PASS result.a is 595098113
PASS result.b is 13480000001
PASS result.a is 615098113
PASS result.b is 13500000001
PASS result.a is 635098113
PASS result.b is 13520000001
PASS result.a is 655098113
PASS result.b is 13540000001
PASS result.a is 675098113
PASS result.b is 13560000001
PASS result.a is 695098113
PASS result.b is 13580000001
PASS result.a is 715098113
PASS result.b is 13600000001
PASS result.a is 735098113
PASS result.b is 13620000001
PASS result.a is 755098113
PASS result.b is 13640000001
PASS result.a is 775098113
PASS result.b is 13660000001
PASS result.a is 795098113
PASS result.b is 13680000001
PASS result.a is 815098113
PASS result.b is 13700000001
PASS result.a is 835098113
PASS result.b is 13720000001
PASS result.a is 855098113
PASS result.b is 13740000001
PASS result.a is 875098113
PASS result.b is 13760000001
PASS result.a is 895098113
PASS result.b is 13780000001
PASS result.a is 915098113
PASS result.b is 13800000001
PASS result.a is 935098113
PASS result.b is 13820000001
PASS result.a is 955098113
PASS result.b is 13840000001
PASS result.a is 975098113
PASS result.b is 13860000001
PASS result.a is 995098113
PASS result.b is 13880000001
PASS result.a is 1015098113
PASS result.b is 13900000001
PASS result.a is 1035098113
PASS result.b is 13920000001
PASS result.a is 1055098113
PASS result.b is 13940000001
PASS result.a is 1075098113
PASS result.b is 13960000001
PASS result.a is 1095098113
PASS result.b is 13980000001
PASS result.a is 1115098113
PASS result.b is 14000000001
PASS result.a is 1135098113
PASS result.b is 14020000001
PASS result.a is 1155098113
PASS result.b is 14040000001
PASS result.a is 1175098113
PASS result.b is 14060000001
PASS result.a is 1195098113
PASS result.b is 14080000001
PASS result.a is 1215098113
PASS result.b is 14100000001
PASS result.a is 1235098113
PASS result.b is 14120000001
PASS result.a is 1255098113
PASS result.b is 14140000001
PASS result.a is 1275098113
PASS result.b is 14160000001
PASS result.a is 1295098113
PASS result.b is 14180000001
PASS result.a is 1315098113
PASS result.b is 14200000001
PASS result.a is 1335098113
PASS result.b is 14220000001
PASS result.a is 1355098113
PASS result.b is 14240000001
PASS result.a is 1375098113
PASS result.b is 14260000001
PASS result.a is 1395098113
PASS result.b is 14280000001
PASS result.a is 1415098113
PASS result.b is 14300000001
PASS result.a is 1435098113
PASS result.b is 14320000001
PASS result.a is 1455098113
PASS result.b is 14340000001
PASS result.a is 1475098113
PASS result.b is 14360000001
PASS result.a is 1495098113
PASS result.b is 14380000001
PASS result.a is 1515098113
PASS result.b is 14400000001
PASS result.a is 1535098113
PASS result.b is 14420000001
PASS result.a is 1555098113
PASS result.b is 14440000001
PASS result.a is 1575098113
PASS result.b is 14460000001
PASS result.a is 1595098113
PASS result.b is 14480000001
PASS result.a is 1615098113
PASS result.b is 14500000001
PASS result.a is 1635098113
PASS result.b is 14520000001
PASS result.a is 1655098113
PASS result.b is 14540000001
PASS result.a is 1675098113
PASS result.b is 14560000001
PASS result.a is 1695098113
PASS result.b is 14580000001
PASS result.a is 1715098113
PASS result.b is 14600000001
PASS result.a is 1735098113
PASS result.b is 14620000001
PASS result.a is 1755098113
PASS result.b is 14640000001
PASS result.a is 1775098113
PASS result.b is 14660000001
PASS result.a is 1795098113
PASS result.b is 14680000001
PASS result.a is 1815098113
PASS result.b is 14700000001
PASS result.a is 1835098113
PASS result.b is 14720000001
PASS result.a is 1855098113
PASS result.b is 14740000001
PASS result.a is 1875098113
PASS result.b is 14760000001
PASS result.a is 1895098113
PASS result.b is 14780000001
PASS result.a is 1915098113
PASS result.b is 14800000001
PASS result.a is 1935098113
PASS result.b is 14820000001
PASS result.a is 1955098113
PASS result.b is 14840000001
PASS result.a is 1975098113
PASS result.b is 14860000001
PASS result.a is 1995098113
PASS result.b is 14880000001
PASS result.a is 2015098113
PASS result.b is 14900000001
PASS result.a is 2035098113
PASS result.b is 14920000001
PASS result.a is 2055098113
PASS result.b is 14940000001
PASS result.a is 2075098113
PASS result.b is 14960000001
PASS result.a is 2095098113
PASS result.b is 14980000001
PASS result.a is 2115098113
PASS result.b is 15000000001
PASS result.a is 2135098113
PASS result.b is 15020000001
PASS result.a is -2139869183
PASS result.b is 15040000001
PASS result.a is -2119869183
PASS result.b is 15060000001
PASS result.a is -2099869183
PASS result.b is 15080000001
PASS result.a is -2079869183
PASS result.b is 15100000001
PASS result.a is -2059869183
PASS result.b is 15120000001
PASS result.a is -2039869183
PASS result.b is 15140000001
PASS result.a is -2019869183
PASS result.b is 15160000001
PASS result.a is -1999869183
PASS result.b is 15180000001
PASS result.a is -1979869183
PASS result.b is 15200000001
PASS result.a is -1959869183
PASS result.b is 15220000001
PASS result.a is -1939869183
PASS result.b is 15240000001
PASS result.a is -1919869183
PASS result.b is 15260000001
PASS result.a is -1899869183
PASS result.b is 15280000001
PASS result.a is -1879869183
PASS result.b is 15300000001
PASS result.a is -1859869183
PASS result.b is 15320000001
PASS result.a is -1839869183
PASS result.b is 15340000001
PASS result.a is -1819869183
PASS result.b is 15360000001
PASS result.a is -1799869183
PASS result.b is 15380000001
PASS result.a is -1779869183
PASS result.b is 15400000001
PASS result.a is -1759869183
PASS result.b is 15420000001
PASS result.a is -1739869183
PASS result.b is 15440000001
PASS result.a is -1719869183
PASS result.b is 15460000001
PASS result.a is -1699869183
PASS result.b is 15480000001
PASS result.a is -1679869183
PASS result.b is 15500000001
PASS result.a is -1659869183
PASS result.b is 15520000001
PASS result.a is -1639869183
PASS result.b is 15540000001
PASS result.a is -1619869183
PASS result.b is 15560000001
PASS result.a is -1599869183
PASS result.b is 15580000001
PASS result.a is -1579869183
PASS result.b is 15600000001
PASS result.a is -1559869183
PASS result.b is 15620000001
PASS result.a is -1539869183
PASS result.b is 15640000001
PASS result.a is -1519869183
PASS result.b is 15660000001
PASS result.a is -1499869183
PASS result.b is 15680000001
PASS result.a is -1479869183
PASS result.b is 15700000001
PASS result.a is -1459869183
PASS result.b is 15720000001
PASS result.a is -1439869183
PASS result.b is 15740000001
PASS result.a is -1419869183
PASS result.b is 15760000001
PASS result.a is -1399869183
PASS result.b is 15780000001
PASS result.a is -1379869183
PASS result.b is 15800000001
PASS result.a is -1359869183
PASS result.b is 15820000001
PASS result.a is -1339869183
PASS result.b is 15840000001
PASS result.a is -1319869183
PASS result.b is 15860000001
PASS result.a is -1299869183
PASS result.b is 15880000001
PASS result.a is -1279869183
PASS result.b is 15900000001
PASS result.a is -1259869183
PASS result.b is 15920000001
PASS result.a is -1239869183
PASS result.b is 15940000001
PASS result.a is -1219869183
PASS result.b is 15960000001
PASS result.a is -1199869183
PASS result.b is 15980000001
PASS result.a is -1179869183
PASS result.b is 16000000001
PASS result.a is -1159869183
PASS result.b is 16020000001
PASS result.a is -1139869183
PASS result.b is 16040000001
PASS result.a is -1119869183
PASS result.b is 16060000001
PASS result.a is -1099869183
PASS result.b is 16080000001
PASS result.a is -1079869183
PASS result.b is 16100000001
PASS result.a is -1059869183
PASS result.b is 16120000001
PASS result.a is -1039869183
PASS result.b is 16140000001
PASS result.a is -1019869183
PASS result.b is 16160000001
PASS result.a is -999869183
PASS result.b is 16180000001
PASS result.a is -979869183
PASS result.b is 16200000001
PASS result.a is -959869183
PASS result.b is 16220000001
PASS result.a is -939869183
PASS result.b is 16240000001
PASS result.a is -919869183
PASS result.b is 16260000001
PASS result.a is -899869183
PASS result.b is 16280000001
PASS result.a is -879869183
PASS result.b is 16300000001
PASS result.a is -859869183
PASS result.b is 16320000001
PASS result.a is -839869183
PASS result.b is 16340000001
PASS result.a is -819869183
PASS result.b is 16360000001
PASS result.a is -799869183
PASS result.b is 16380000001
PASS result.a is -779869183
PASS result.b is 16400000001
PASS result.a is -759869183
PASS result.b is 16420000001
PASS result.a is -739869183
PASS result.b is 16440000001
PASS result.a is -719869183
PASS result.b is 16460000001
PASS result.a is -699869183
PASS result.b is 16480000001
PASS result.a is -679869183
PASS result.b is 16500000001
PASS result.a is -659869183
PASS result.b is 16520000001
PASS result.a is -639869183
PASS result.b is 16540000001
PASS result.a is -619869183
PASS result.b is 16560000001
PASS result.a is -599869183
PASS result.b is 16580000001
PASS result.a is -579869183
PASS result.b is 16600000001
PASS result.a is -559869183
PASS result.b is 16620000001
PASS result.a is -539869183
PASS result.b is 16640000001
PASS result.a is -519869183
PASS result.b is 16660000001
PASS result.a is -499869183
PASS result.b is 16680000001
PASS result.a is -479869183
PASS result.b is 16700000001
PASS result.a is -459869183
PASS result.b is 16720000001
PASS result.a is -439869183
PASS result.b is 16740000001
PASS result.a is -419869183
PASS result.b is 16760000001
PASS result.a is -399869183
PASS result.b is 16780000001
PASS result.a is -379869183
PASS result.b is 16800000001
PASS result.a is -359869183
PASS result.b is 16820000001
PASS result.a is -339869183
PASS result.b is 16840000001
PASS result.a is -319869183
PASS result.b is 16860000001
PASS result.a is -299869183
PASS result.b is 16880000001
PASS result.a is -279869183
PASS result.b is 16900000001
PASS result.a is -259869183
PASS result.b is 16920000001
PASS result.a is -239869183
PASS result.b is 16940000001
PASS result.a is -219869183
PASS result.b is 16960000001
PASS result.a is -199869183
PASS result.b is 16980000001
PASS result.a is -179869183
PASS result.b is 17000000001
PASS result.a is -159869183
PASS result.b is 17020000001
PASS result.a is -139869183
PASS result.b is 17040000001
PASS result.a is -119869183
PASS result.b is 17060000001
PASS result.a is -99869183
PASS result.b is 17080000001
PASS result.a is -79869183
PASS result.b is 17100000001
PASS result.a is -59869183
PASS result.b is 17120000001
PASS result.a is -39869183
PASS result.b is 17140000001
PASS result.a is -19869183
PASS result.b is 17160000001
PASS result.a is 130817
PASS result.b is 17180000001
PASS result.a is 20130817
PASS result.b is 17200000001
PASS result.a is 40130817
PASS result.b is 17220000001
PASS result.a is 60130817
PASS result.b is 17240000001
PASS result.a is 80130817
PASS result.b is 17260000001
PASS result.a is 100130817
PASS result.b is 17280000001
PASS result.a is 120130817
PASS result.b is 17300000001
PASS result.a is 140130817
PASS result.b is 17320000001
PASS result.a is 160130817
PASS result.b is 17340000001
PASS result.a is 180130817
PASS result.b is 17360000001
PASS result.a is 200130817
PASS result.b is 17380000001
PASS result.a is 220130817
PASS result.b is 17400000001
PASS result.a is 240130817
PASS result.b is 17420000001
PASS result.a is 260130817
PASS result.b is 17440000001
PASS result.a is 280130817
PASS result.b is 17460000001
PASS result.a is 300130817
PASS result.b is 17480000001
PASS result.a is 320130817
PASS result.b is 17500000001
PASS result.a is 340130817
PASS result.b is 17520000001
PASS result.a is 360130817
PASS result.b is 17540000001
PASS result.a is 380130817
PASS result.b is 17560000001
PASS result.a is 400130817
PASS result.b is 17580000001
PASS result.a is 420130817
PASS result.b is 17600000001
PASS result.a is 440130817
PASS result.b is 17620000001
PASS result.a is 460130817
PASS result.b is 17640000001
PASS result.a is 480130817
PASS result.b is 17660000001
PASS result.a is 500130817
PASS result.b is 17680000001
PASS result.a is 520130817
PASS result.b is 17700000001
PASS result.a is 540130817
PASS result.b is 17720000001
PASS result.a is 560130817
PASS result.b is 17740000001
PASS result.a is 580130817
PASS result.b is 17760000001
PASS result.a is 600130817
PASS result.b is 17780000001
PASS result.a is 620130817
PASS result.b is 17800000001
PASS result.a is 640130817
PASS result.b is 17820000001
PASS result.a is 660130817
PASS result.b is 17840000001
PASS result.a is 680130817
PASS result.b is 17860000001
PASS result.a is 700130817
PASS result.b is 17880000001
PASS result.a is 720130817
PASS result.b is 17900000001
PASS result.a is 740130817
PASS result.b is 17920000001
PASS result.a is 760130817
PASS result.b is 17940000001
PASS result.a is 780130817
PASS result.b is 17960000001
PASS result.a is 800130817
PASS result.b is 17980000001
PASS result.a is 820130817
PASS result.b is 18000000001
PASS result.a is 840130817
PASS result.b is 18020000001
PASS result.a is 860130817
PASS result.b is 18040000001
PASS result.a is 880130817
PASS result.b is 18060000001
PASS result.a is 900130817
PASS result.b is 18080000001
PASS result.a is 920130817
PASS result.b is 18100000001
PASS result.a is 940130817
PASS result.b is 18120000001
PASS result.a is 960130817
PASS result.b is 18140000001
PASS result.a is 980130817
PASS result.b is 18160000001
PASS result.a is 1000130817
PASS result.b is 18180000001
PASS result.a is 1020130817
PASS result.b is 18200000001
PASS result.a is 1040130817
PASS result.b is 18220000001
PASS result.a is 1060130817
PASS result.b is 18240000001
PASS result.a is 1080130817
PASS result.b is 18260000001
PASS result.a is 1100130817
PASS result.b is 18280000001
PASS result.a is 1120130817
PASS result.b is 18300000001
PASS result.a is 1140130817
PASS result.b is 18320000001
PASS result.a is 1160130817
PASS result.b is 18340000001
PASS result.a is 1180130817
PASS result.b is 18360000001
PASS result.a is 1200130817
PASS result.b is 18380000001
PASS result.a is 1220130817
PASS result.b is 18400000001
PASS result.a is 1240130817
PASS result.b is 18420000001
PASS result.a is 1260130817
PASS result.b is 18440000001
PASS result.a is 1280130817
PASS result.b is 18460000001
PASS result.a is 1300130817
PASS result.b is 18480000001
PASS result.a is 1320130817
PASS result.b is 18500000001
PASS result.a is 1340130817
PASS result.b is 18520000001
PASS result.a is 1360130817
PASS result.b is 18540000001
PASS result.a is 1380130817
PASS result.b is 18560000001
PASS result.a is 1400130817
PASS result.b is 18580000001
PASS result.a is 1420130817
PASS result.b is 18600000001
PASS result.a is 1440130817
PASS result.b is 18620000001
PASS result.a is 1460130817
PASS result.b is 18640000001
PASS result.a is 1480130817
PASS result.b is 18660000001
PASS result.a is 1500130817
PASS result.b is 18680000001
PASS result.a is 1520130817
PASS result.b is 18700000001
PASS result.a is 1540130817
PASS result.b is 18720000001
PASS result.a is 1560130817
PASS result.b is 18740000001
PASS result.a is 1580130817
PASS result.b is 18760000001
PASS result.a is 1600130817
PASS result.b is 18780000001
PASS result.a is 1620130817
PASS result.b is 18800000001
PASS result.a is 1640130817
PASS result.b is 18820000001
PASS result.a is 1660130817
PASS result.b is 18840000001
PASS result.a is 1680130817
PASS result.b is 18860000001
PASS result.a is 1700130817
PASS result.b is 18880000001
PASS result.a is 1720130817
PASS result.b is 18900000001
PASS result.a is 1740130817
PASS result.b is 18920000001
PASS result.a is 1760130817
PASS result.b is 18940000001
PASS result.a is 1780130817
PASS result.b is 18960000001
PASS result.a is 1800130817
PASS result.b is 18980000001
PASS result.a is 1820130817
PASS result.b is 19000000001
PASS result.a is 1840130817
PASS result.b is 19020000001
PASS result.a is 1860130817
PASS result.b is 19040000001
PASS result.a is 1880130817
PASS result.b is 19060000001
PASS result.a is 1900130817
PASS result.b is 19080000001
PASS result.a is 1920130817
PASS result.b is 19100000001
PASS result.a is 1940130817
PASS result.b is 19120000001
PASS result.a is 1960130817
PASS result.b is 19140000001
PASS result.a is 1980130817
PASS result.b is 19160000001
PASS result.a is 2000130817
PASS result.b is 19180000001
PASS result.a is 2020130817
PASS result.b is 19200000001
PASS result.a is 2040130817
PASS result.b is 19220000001
PASS result.a is 2060130817
PASS result.b is 19240000001
PASS result.a is 2080130817
PASS result.b is 19260000001
PASS result.a is 2100130817
PASS result.b is 19280000001
PASS result.a is 2120130817
PASS result.b is 19300000001
PASS result.a is 2140130817
PASS result.b is 19320000001
PASS result.a is -2134836479
PASS result.b is 19340000001
PASS result.a is -2114836479
PASS result.b is 19360000001
PASS result.a is -2094836479
PASS result.b is 19380000001
PASS result.a is -2074836479
PASS result.b is 19400000001
PASS result.a is -2054836479
PASS result.b is 19420000001
PASS result.a is -2034836479
PASS result.b is 19440000001
PASS result.a is -2014836479
PASS result.b is 19460000001
PASS result.a is -1994836479
PASS result.b is 19480000001
PASS result.a is -1974836479
PASS result.b is 19500000001
PASS result.a is -1954836479
PASS result.b is 19520000001
PASS result.a is -1934836479
PASS result.b is 19540000001
PASS result.a is -1914836479
PASS result.b is 19560000001
PASS result.a is -1894836479
PASS result.b is 19580000001
PASS result.a is -1874836479
PASS result.b is 19600000001
PASS result.a is -1854836479
PASS result.b is 19620000001
PASS result.a is -1834836479
PASS result.b is 19640000001
PASS result.a is -1814836479
PASS result.b is 19660000001
PASS result.a is -1794836479
PASS result.b is 19680000001
PASS result.a is -1774836479
PASS result.b is 19700000001
PASS result.a is -1754836479
PASS result.b is 19720000001
PASS result.a is -1734836479
PASS result.b is 19740000001
PASS result.a is -1714836479
PASS result.b is 19760000001
PASS result.a is -1694836479
PASS result.b is 19780000001
PASS result.a is -1674836479
PASS result.b is 19800000001
PASS result.a is -1654836479
PASS result.b is 19820000001
PASS result.a is -1634836479
PASS result.b is 19840000001
PASS result.a is -1614836479
PASS result.b is 19860000001
PASS result.a is -1594836479
PASS result.b is 19880000001
PASS result.a is -1574836479
PASS result.b is 19900000001
PASS result.a is -1554836479
PASS result.b is 19920000001
PASS result.a is -1534836479
PASS result.b is 19940000001
PASS result.a is -1514836479
PASS result.b is 19960000001
PASS result.a is -1494836479
PASS result.b is 19980000001
PASS successfullyParsed is true

TEST COMPLETE
